requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
schedule==1.2.0
Werkzeug==2.3.7